+++ Guardfile
@@ -13,6 +13,6 @@ guard :minitest do
 end
 
 guard 'rake', task: 'build_parser' do
-  watch("lib/graphql/language/parse.y")
+  watch("lib/graphql/language/parser.y")
   watch("lib/graphql/language/lexer.rl")
 end
