+++ Rakefile
@@ -33,8 +33,9 @@ task :console do
   IRB.start
 end
 
+desc "Use Racc & Ragel to regenerate parser.rb & lexer.rb from configuration files"
 task :build_parser do
-  `rm lib/graphql/language/racc_parser.rb lib/graphql/language/lexer.rb `
-  `racc lib/graphql/language/parse.y -o lib/graphql/language/racc_parser.rb`
+  `rm lib/graphql/language/parser.rb lib/graphql/language/lexer.rb `
+  `racc lib/graphql/language/parser.y -o lib/graphql/language/parser.rb`
   `ragel -R lib/graphql/language/lexer.rl`
 end
