+++ lib/graphql/language.rb
@@ -1,12 +1,5 @@
-require 'graphql/language/parser'
-require 'graphql/language/transform'
-require 'graphql/language/nodes'
-require 'graphql/language/visitor'
-require 'graphql/language/racc_parser'
 require 'graphql/language/lexer'
+require 'graphql/language/nodes'
+require 'graphql/language/parser'
 require 'graphql/language/token'
-
-module GraphQL
-  TRANSFORM = GraphQL::Language::Transform.new
-  PARSER = GraphQL::Language::Parser.new
-end
+require 'graphql/language/visitor'
