+++ lib/graphql/language/parser.rb
@@ -1,122 +1,811 @@
+#
+# DO NOT MODIFY!!!!
+# This file is automatically generated by Racc 1.4.14
+# from Racc grammer file "".
+#
+
+require 'racc/parser.rb'
+
+
+
 module GraphQL
   module Language
-    # Parser is a [parslet](http://kschiess.github.io/parslet/) parser for parsing queries.
-    #
-    class Parser < Parslet::Parser
-      root(:document)
-      rule(:document) { (
-          space                |
-          operation_definition |
-          fragment_definition
-        ).repeat.as(:document_parts)
-      }
+    class Parser < Racc::Parser
 
-      # TODO: whitespace sensitive regarding `on`, eg `onFood`, see lookahead note in spec
-      rule(:fragment_definition) {
-        str("fragment").as(:fragment_keyword) >>
-         space? >> name.as(:fragment_name) >>
-         space? >> str("on") >> space? >> name.as(:type_condition) >>
-         space? >> directives.maybe.as(:optional_directives).as(:directives) >>
-         space? >> selections.as(:selections)
-      }
+module_eval(<<'...end parser.y/module_eval...', 'parser.y', 200)
 
-      rule(:fragment_spread) {
-        spread.as(:fragment_spread_keyword) >> space? >>
-        name.as(:fragment_spread_name) >> space? >>
-        directives.maybe.as(:optional_directives).as(:directives)
-      }
-      rule(:spread) { str("...") }
-      rule(:type_condition) { str("on ") >> name.as(:optional_string_content) }
-      # TODO: `on` bug, see spec
-      rule(:inline_fragment) {
-        spread.as(:fragment_spread_keyword) >> space? >>
-        type_condition.maybe.as(:inline_fragment_type) >> space? >>
-        directives.maybe.as(:optional_directives).as(:directives) >> space? >>
-        selections.as(:selections)
-      }
+def initialize(query_string)
+  @query_string = query_string
+end
 
-      rule(:operation_definition) { (unnamed_selections | typed_operation_definition) }
-      rule(:unnamed_selections) { selections.as(:unnamed_selections)}
-      rule(:typed_operation_definition) {
-        operation_type.as(:operation_type) >> space? >>
-        name.maybe.as(:name) >> space? >>
-        operation_variable_definitions.maybe.as(:optional_variables).as(:variables) >> space? >>
-        directives.maybe.as(:optional_directives).as(:directives) >> space? >>
-        selections.as(:selections)
-      }
-      rule(:operation_type) { (str("query") | str("mutation") | str("subscription")) }
-      rule(:operation_variable_definitions) { str("(") >> space? >> (operation_variable_definition >> separator?).repeat(1) >> space? >> str(")") }
-      rule(:operation_variable_definition) {
-        value_variable.as(:variable_name) >> space? >>
-        str(":") >> space? >>
-        type.as(:variable_type) >> space? >>
-        (str("=") >> space? >> value.as(:variable_default_value)).maybe.as(:variable_optional_default_value)
-      }
-
-      rule(:selection) { (inline_fragment | fragment_spread | field) >> space? >> separator? }
-      rule(:selections) { str("{") >> space? >> selection.repeat(1) >> space? >> str("}")}
-
-      rule(:field) {
-        field_alias.maybe.as(:alias) >>
-        name.as(:field_name) >>
-        field_arguments.maybe.as(:optional_field_arguments).as(:field_arguments) >> space? >>
-        directives.maybe.as(:optional_directives).as(:directives) >> space? >>
-        selections.maybe.as(:optional_selections).as(:selections)
-      }
-
-      rule(:field_alias) { name.as(:alias_name) >> space? >> str(":") >> space? }
-      rule(:field_arguments) { str("(") >> space? >> field_argument.repeat(1) >> space? >> str(")") }
-      rule(:field_argument) { name.as(:field_argument_name) >> str(":") >> space? >> value.as(:field_argument_value) >> separator? }
-
-      rule(:directives) { (directive >> separator?).repeat(1) }
-      rule(:directive) {
-        str("@") >> name.as(:directive_name) >>
-        directive_arguments.maybe.as(:optional_directive_arguments).as(:directive_arguments)
-      }
-      rule(:directive_arguments) { str("(") >> directive_argument.repeat(1) >> str(")") }
-      rule(:directive_argument) { name.as(:directive_argument_name) >> str(":") >> space? >> value.as(:directive_argument_value) >> separator? }
-
-      rule(:type) { (non_null_type | list_type | type_name)}
-      rule(:list_type) { str("[") >> type.as(:list_type) >> str("]")}
-      rule(:non_null_type) { (list_type | type_name).as(:non_null_type) >> str("!") }
-      rule(:type_name) { name.as(:type_name) }
-
-      rule(:value) {(
-        value_input_object  |
-        value_float         |
-        value_int           |
-        value_string        |
-        value_boolean       |
-        value_array         |
-        value_variable      |
-        value_enum
-      )}
-      rule(:value_sign?) { match('[\-\+]').maybe }
-      rule(:value_array) { (str("[") >> space? >> (value >> separator?).repeat(0) >> str("]")).as(:array) }
-      rule(:value_boolean) { (str("true") | str("false")).as(:boolean) }
-      rule(:value_float) { (value_sign? >> match('\d').repeat(1) >> str(".") >> match('\d').repeat(1) >> (match("[eE]") >> value_sign? >> match('\d').repeat(1)).maybe).as(:float) }
-      rule(:value_input_object) { str("{") >> space? >> value_input_object_pair.repeat(0).as(:input_object) >> space? >> str("}") }
-      rule(:value_input_object_pair) { space? >> name.as(:input_object_name) >>  space? >> str(":") >> space? >> value.as(:input_object_value) >> separator? }
-      rule(:value_int) { (value_sign? >> match('\d').repeat(1)).as(:int) }
-      rule(:value_string) { str('"') >> value_string_char.repeat.maybe.as(:optional_string_content).as(:string) >> str('"')}
-      rule(:value_string_char) { value_string_escaped_char | value_string_escaped_unicode | value_string_source_char}
-      rule(:value_string_escaped_char) { str("\\") >> match('["\\\\/bfnrt]') }
-      rule(:value_string_escaped_unicode) { str("\\") >> match('u[\dA-Fa-f]{4}')}
-      rule(:value_string_source_char) { (str('"') | str("\\") | value_string_line_terminator).absent? >> any }
-      rule(:value_string_line_terminator) {
-        str('\u000A') | # new line
-        str('\u000D') | # carriage return
-        str('\u2028') | # line separator
-        str('\u2029')   # paragraph separator
-      }
-      rule(:value_enum) { name.as(:enum) }
-      rule(:value_variable) { str("$") >> name.as(:variable) }
-
-      rule(:separator?) { space? >> str(",").maybe >> space? }
-      rule(:name) { match('[_A-Za-z]') >> match('[_0-9A-Za-z]').repeat(0) }
-      rule(:comment) { str("#") >> match('[^\r\n]').repeat(0) }
-      rule(:space) { (match('[\s\n]+') | comment).repeat(1) }
-      rule(:space?) { space.maybe }
+def parse_document
+  @document ||= begin
+    @tokens ||= GraphQL::Language::Lexer.tokenize(@query_string)
+    if @tokens.none?
+      make_node(:Document, definitions: [])
+    else
+      do_parse
     end
   end
 end
+
+def self.parse(query_string)
+  self.new(query_string).parse_document
+end
+
+private
+
+def next_token
+  lexer_token = @tokens.shift
+  if lexer_token.nil?
+    nil
+  else
+    [lexer_token.name, lexer_token]
+  end
+end
+
+def on_error(parser_token_id, lexer_token, vstack)
+  if lexer_token == "$"
+    raise GraphQL::ParseError.new("Unexpected end of document", nil, nil, @query_string)
+  else
+    parser_token_name = token_to_str(parser_token_id)
+    line, col = lexer_token.line_and_column
+    raise GraphQL::ParseError.new("Parse error on #{lexer_token.to_s.inspect} (#{parser_token_name}) at [#{line}, #{col}]", line, col, @query_string)
+  end
+end
+
+def make_node(node_name, assigns)
+  assigns.each do |key, value|
+    if key != :position_source && value.is_a?(GraphQL::Language::Token)
+      assigns[key] = value.to_s
+    end
+  end
+
+  GraphQL::Language::Nodes.const_get(node_name).new(assigns)
+end
+...end parser.y/module_eval...
+##### State transition tables begin ###
+
+racc_action_table = [
+    81,   100,    35,    82,   101,    34,    83,   100,    84,   109,
+    75,    76,    35,    72,    73,    74,    81,    57,    45,    82,
+    42,    42,    83,    65,    84,    69,    75,    76,    42,    72,
+    73,    74,    81,    35,     9,    82,     9,    30,    83,     9,
+    84,    38,    75,    76,    28,    72,    73,    74,    81,    42,
+     9,    82,   103,    42,    83,    42,    84,    42,    75,    76,
+   107,    72,    73,    74,    81,    15,    45,    82,    92,     9,
+    83,    63,    84,   nil,    75,    76,   nil,    72,    73,    74,
+    10,    19,    12,    13,    14,    31,    10,    19,    12,    13,
+    14,   nil,   nil,   nil,   nil,    26,    10,    19,    12,    13,
+    14,    89,   nil,   nil,   nil,    26,    10,    19,    12,    13,
+    14,    89,   nil,   nil,   nil,   nil,    10,    19,    12,    13,
+    14,     9,   nil,    10,    11,    12,    13,    14,     9,   nil,
+    10,    11,    12,    13,    14,   105,    10,    19,    12,    13,
+    14,    95,    10,    19,    12,    13,    14,    10,    19,    12,
+    13,    14,    10,    19,    12,    13,    14,    10,    19,    12,
+    13,    14,    10,    19,    12,    13,    14,    10,    19,    12,
+    13,    14,    10,    19,    12,    13,    14,    10,    19,    12,
+    13,    14,    10,    19,    12,    13,    14,    10,    19,    12,
+    13,    14,    10,    19,    12,    13,    36 ]
+
+racc_action_check = [
+   107,    87,    25,   107,    87,    25,   107,   102,   107,   102,
+   107,   107,    56,   107,   107,   107,   101,    43,    43,   101,
+    51,    53,   101,    50,   101,    59,   101,   101,    62,   101,
+   101,   101,    65,    47,    66,    65,    67,    17,    65,    70,
+    65,    27,    65,    65,    15,    65,    65,    65,    93,    40,
+    39,    93,    93,    37,    93,    29,    93,    33,    93,    93,
+    98,    93,    93,    93,    82,     1,    30,    82,    82,    46,
+    82,    48,    82,   nil,    82,    82,   nil,    82,    82,    82,
+    48,    48,    48,    48,    48,    20,    20,    20,    20,    20,
+    20,   nil,   nil,   nil,   nil,    20,     9,     9,     9,     9,
+     9,    69,   nil,   nil,   nil,     9,    69,    69,    69,    69,
+    69,    89,   nil,   nil,   nil,   nil,    89,    89,    89,    89,
+    89,     3,   nil,     3,     3,     3,     3,     3,     0,   nil,
+     0,     0,     0,     0,     0,    96,    96,    96,    96,    96,
+    96,    83,    83,    83,    83,    83,    83,    36,    36,    36,
+    36,    36,     7,     7,     7,     7,     7,    35,    35,    35,
+    35,    35,    34,    34,    34,    34,    34,    11,    11,    11,
+    11,    11,    38,    38,    38,    38,    38,    45,    45,    45,
+    45,    45,    81,    81,    81,    81,    81,    42,    42,    42,
+    42,    42,    26,    26,    26,    26,    26 ]
+
+racc_action_pointer = [
+   118,    65,   nil,   111,   nil,   nil,   nil,   140,   nil,    84,
+   nil,   155,   nil,   nil,   nil,    44,   nil,    35,   nil,   nil,
+    74,   nil,   nil,   nil,   nil,     0,   180,    25,   nil,    35,
+    62,   nil,   nil,    37,   150,   145,   135,    33,   160,    40,
+    29,   nil,   175,    14,   nil,   165,    59,    31,    68,   nil,
+    18,     0,   nil,     1,   nil,   nil,    10,   nil,   nil,    20,
+   nil,   nil,     8,   nil,   nil,    28,    24,    26,   nil,    94,
+    29,   nil,   nil,   nil,   nil,   nil,   nil,   nil,   nil,   nil,
+   nil,   170,    60,   130,   nil,   nil,   nil,    -5,   nil,   104,
+   nil,   nil,   nil,    44,   nil,   nil,   124,   nil,    55,   nil,
+   nil,    12,     1,   nil,   nil,   nil,   nil,    -4,   nil,   nil,
+   nil ]
+
+racc_action_default = [
+   -69,   -69,    -1,    -2,    -3,    -5,    -6,    -9,    -8,   -69,
+   -31,   -32,   -33,   -34,   -35,   -69,    -4,   -11,   -10,   -32,
+   -69,   -24,   -26,   -27,   -28,   -36,   -69,   -69,   111,   -61,
+   -69,   -21,   -25,   -61,   -69,   -69,   -35,   -61,   -69,   -69,
+   -62,   -63,   -69,   -69,   -13,   -69,   -22,   -36,   -69,   -38,
+   -69,   -61,   -66,   -61,    -7,   -64,   -36,   -12,   -14,   -69,
+   -23,   -29,   -61,   -37,   -39,   -69,   -69,   -69,   -65,   -69,
+   -22,   -40,   -41,   -42,   -43,   -44,   -45,   -46,   -47,   -48,
+   -49,   -69,   -69,   -69,   -60,   -67,   -68,   -19,   -16,   -69,
+   -30,   -50,   -51,   -69,   -53,   -55,   -69,   -57,   -69,   -15,
+   -17,   -69,   -69,   -52,   -54,   -56,   -58,   -69,   -20,   -18,
+   -59 ]
+
+racc_goto_table = [
+    18,    61,    25,    97,    27,    54,    71,    49,    44,    99,
+    43,     4,    60,    25,    16,    20,   106,    87,    33,    37,
+    64,    58,    39,    94,    29,    90,    46,    47,    50,    51,
+    52,    53,    85,    86,   104,    56,    60,   102,    59,    21,
+    62,    50,   108,    17,    66,     1,    67,    48,   110,    68,
+    32,     3,    93,    96,     2,    70,    55,   nil,   nil,   nil,
+   nil,   nil,    88,   nil,   nil,   nil,   nil,   nil,   nil,   nil,
+   nil,   nil,   nil,   nil,    91,   nil,    98,   nil,   nil,   nil,
+   nil,   nil,    88,   nil,   nil,   nil,   nil,   nil,   nil,    98 ]
+
+racc_goto_check = [
+     7,    18,     7,    32,     7,    11,    16,    25,    13,    15,
+    12,     4,    11,     7,     4,    17,    32,    14,    23,     7,
+    25,    13,    10,    16,     9,    18,    10,     7,     7,     7,
+    10,     7,    11,    11,    16,     7,    11,    14,     7,    19,
+    23,     7,    16,     8,    10,     1,    10,    24,    16,    23,
+    19,     3,    30,    31,     2,    10,    34,   nil,   nil,   nil,
+   nil,   nil,     7,   nil,   nil,   nil,   nil,   nil,   nil,   nil,
+   nil,   nil,   nil,   nil,     7,   nil,     7,   nil,   nil,   nil,
+   nil,   nil,     7,   nil,   nil,   nil,   nil,   nil,   nil,     7 ]
+
+racc_goto_pointer = [
+   nil,    45,    54,    51,    11,   nil,   nil,    -7,    36,     7,
+    -7,   -34,   -20,   -22,   -52,   -78,   -59,     6,   -45,    30,
+   nil,   nil,   nil,    -7,    12,   -28,   nil,   nil,   nil,   nil,
+   -30,   -30,   -80,   nil,    16 ]
+
+racc_goto_default = [
+   nil,   nil,   nil,   nil,   nil,     5,     6,     7,   nil,   nil,
+   nil,     8,   nil,   nil,   nil,   nil,   nil,   nil,   nil,   nil,
+    22,    23,    24,   nil,   nil,   nil,    77,    78,    79,    80,
+   nil,   nil,   nil,    40,    41 ]
+
+racc_reduce_table = [
+  0, 0, :racc_error,
+  1, 23, :_reduce_none,
+  1, 24, :_reduce_2,
+  1, 25, :_reduce_3,
+  2, 25, :_reduce_4,
+  1, 26, :_reduce_none,
+  1, 26, :_reduce_none,
+  5, 27, :_reduce_7,
+  1, 27, :_reduce_8,
+  0, 30, :_reduce_9,
+  1, 30, :_reduce_none,
+  0, 31, :_reduce_11,
+  3, 31, :_reduce_12,
+  1, 34, :_reduce_13,
+  2, 34, :_reduce_14,
+  5, 35, :_reduce_15,
+  1, 36, :_reduce_16,
+  2, 36, :_reduce_17,
+  3, 36, :_reduce_18,
+  0, 37, :_reduce_19,
+  2, 37, :_reduce_20,
+  3, 33, :_reduce_21,
+  0, 40, :_reduce_22,
+  1, 40, :_reduce_23,
+  1, 39, :_reduce_24,
+  2, 39, :_reduce_25,
+  1, 41, :_reduce_none,
+  1, 41, :_reduce_none,
+  1, 41, :_reduce_none,
+  4, 42, :_reduce_29,
+  6, 42, :_reduce_30,
+  1, 29, :_reduce_none,
+  1, 29, :_reduce_none,
+  1, 29, :_reduce_none,
+  1, 29, :_reduce_none,
+  1, 29, :_reduce_none,
+  0, 45, :_reduce_36,
+  3, 45, :_reduce_37,
+  1, 46, :_reduce_38,
+  2, 46, :_reduce_39,
+  3, 47, :_reduce_40,
+  1, 38, :_reduce_41,
+  1, 38, :_reduce_42,
+  1, 38, :_reduce_43,
+  1, 38, :_reduce_44,
+  1, 38, :_reduce_45,
+  1, 38, :_reduce_none,
+  1, 38, :_reduce_none,
+  1, 38, :_reduce_none,
+  1, 38, :_reduce_none,
+  2, 48, :_reduce_50,
+  2, 49, :_reduce_51,
+  3, 49, :_reduce_52,
+  1, 52, :_reduce_53,
+  2, 52, :_reduce_54,
+  2, 50, :_reduce_55,
+  3, 50, :_reduce_56,
+  1, 53, :_reduce_57,
+  2, 53, :_reduce_58,
+  3, 54, :_reduce_59,
+  1, 51, :_reduce_60,
+  0, 32, :_reduce_61,
+  1, 32, :_reduce_none,
+  1, 55, :_reduce_63,
+  2, 55, :_reduce_64,
+  3, 56, :_reduce_65,
+  3, 43, :_reduce_66,
+  5, 44, :_reduce_67,
+  6, 28, :_reduce_68 ]
+
+racc_reduce_n = 69
+
+racc_shift_n = 111
+
+racc_token_table = {
+  false => 0,
+  :error => 1,
+  :RPAREN => 2,
+  :LPAREN => 3,
+  :VAR_SIGN => 4,
+  :COLON => 5,
+  :BANG => 6,
+  :RBRACKET => 7,
+  :LBRACKET => 8,
+  :EQUALS => 9,
+  :RCURLY => 10,
+  :LCURLY => 11,
+  :IDENTIFIER => 12,
+  :FRAGMENT => 13,
+  :TRUE => 14,
+  :FALSE => 15,
+  :ON => 16,
+  :FLOAT => 17,
+  :INT => 18,
+  :STRING => 19,
+  :DIR_SIGN => 20,
+  :ELLIPSIS => 21 }
+
+racc_nt_base = 22
+
+racc_use_result_var = true
+
+Racc_arg = [
+  racc_action_table,
+  racc_action_check,
+  racc_action_default,
+  racc_action_pointer,
+  racc_goto_table,
+  racc_goto_check,
+  racc_goto_default,
+  racc_goto_pointer,
+  racc_nt_base,
+  racc_reduce_table,
+  racc_token_table,
+  racc_shift_n,
+  racc_reduce_n,
+  racc_use_result_var ]
+
+Racc_token_to_s_table = [
+  "$end",
+  "error",
+  "RPAREN",
+  "LPAREN",
+  "VAR_SIGN",
+  "COLON",
+  "BANG",
+  "RBRACKET",
+  "LBRACKET",
+  "EQUALS",
+  "RCURLY",
+  "LCURLY",
+  "IDENTIFIER",
+  "FRAGMENT",
+  "TRUE",
+  "FALSE",
+  "ON",
+  "FLOAT",
+  "INT",
+  "STRING",
+  "DIR_SIGN",
+  "ELLIPSIS",
+  "$start",
+  "target",
+  "document",
+  "definitions_list",
+  "definition",
+  "operation_definition",
+  "fragment_definition",
+  "name",
+  "operation_name_opt",
+  "variable_definitions_opt",
+  "directives_list_opt",
+  "selection_set",
+  "variable_definitions_list",
+  "variable_definition",
+  "variable_definition_type_name",
+  "variable_definition_default_value_opt",
+  "input_value",
+  "selection_list",
+  "selection_set_opt",
+  "selection",
+  "field",
+  "fragment_spread",
+  "inline_fragment",
+  "arguments_opt",
+  "arguments_list",
+  "argument",
+  "variable",
+  "list_value",
+  "object_value",
+  "enum_value",
+  "list_value_list",
+  "object_value_list",
+  "object_value_field",
+  "directives_list",
+  "directive" ]
+
+Racc_debug_parser = false
+
+##### State transition tables end #####
+
+# reduce 0 omitted
+
+# reduce 1 omitted
+
+module_eval(<<'.,.,', 'parser.y', 4)
+  def _reduce_2(val, _values, result)
+     return make_node(:Document, definitions: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 7)
+  def _reduce_3(val, _values, result)
+     return [val[0]]
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 8)
+  def _reduce_4(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+# reduce 5 omitted
+
+# reduce 6 omitted
+
+module_eval(<<'.,.,', 'parser.y', 16)
+  def _reduce_7(val, _values, result)
+            return  make_node(
+          :OperationDefinition, {
+            operation_type: val[0],
+            name:           val[1],
+            variables:      val[2],
+            directives:     val[3],
+            selections:     val[4],
+            position_source: val[0],
+          }
+        )
+      
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 28)
+  def _reduce_8(val, _values, result)
+            return make_node(
+          :OperationDefinition, {
+            operation_type: "query",
+            selections: val[0],
+          }
+        )
+      
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 37)
+  def _reduce_9(val, _values, result)
+     return nil 
+    result
+  end
+.,.,
+
+# reduce 10 omitted
+
+module_eval(<<'.,.,', 'parser.y', 41)
+  def _reduce_11(val, _values, result)
+     return [] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 42)
+  def _reduce_12(val, _values, result)
+     return val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 45)
+  def _reduce_13(val, _values, result)
+     return [val[0]] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 46)
+  def _reduce_14(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 50)
+  def _reduce_15(val, _values, result)
+            return make_node(:VariableDefinition, {
+          name: val[1],
+          type: val[3],
+          default_value: val[4],
+          position_source: val[0],
+        })
+      
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 59)
+  def _reduce_16(val, _values, result)
+     return make_node(:TypeName, name: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 60)
+  def _reduce_17(val, _values, result)
+     return make_node(:NonNullType, of_type: val[0]) 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 61)
+  def _reduce_18(val, _values, result)
+     return make_node(:ListType, of_type: val[1]) 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 64)
+  def _reduce_19(val, _values, result)
+     return nil 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 65)
+  def _reduce_20(val, _values, result)
+     return val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 67)
+  def _reduce_21(val, _values, result)
+     return val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 70)
+  def _reduce_22(val, _values, result)
+     return [] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 71)
+  def _reduce_23(val, _values, result)
+     return val[0] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 74)
+  def _reduce_24(val, _values, result)
+     return [result] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 75)
+  def _reduce_25(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+# reduce 26 omitted
+
+# reduce 27 omitted
+
+# reduce 28 omitted
+
+module_eval(<<'.,.,', 'parser.y', 84)
+  def _reduce_29(val, _values, result)
+                return make_node(
+              :Field, {
+                name:         val[0],
+                arguments:    val[1],
+                directives:   val[2],
+                selections:   val[3],
+                position_source: val[0],
+              }
+            )
+          
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 95)
+  def _reduce_30(val, _values, result)
+                return make_node(
+              :Field, {
+                alias:        val[0],
+                name:         val[2],
+                arguments:    val[3],
+                directives:   val[4],
+                selections:   val[5],
+                position_source: val[0],
+              }
+            )
+          
+    result
+  end
+.,.,
+
+# reduce 31 omitted
+
+# reduce 32 omitted
+
+# reduce 33 omitted
+
+# reduce 34 omitted
+
+# reduce 35 omitted
+
+module_eval(<<'.,.,', 'parser.y', 115)
+  def _reduce_36(val, _values, result)
+     return [] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 116)
+  def _reduce_37(val, _values, result)
+     return val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 119)
+  def _reduce_38(val, _values, result)
+     return [val[0]] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 120)
+  def _reduce_39(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 123)
+  def _reduce_40(val, _values, result)
+     return make_node(:Argument, name: val[0], value: val[2], position_source: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 126)
+  def _reduce_41(val, _values, result)
+     return val[0].to_f 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 127)
+  def _reduce_42(val, _values, result)
+     return val[0].to_i 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 128)
+  def _reduce_43(val, _values, result)
+     return val[0] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 129)
+  def _reduce_44(val, _values, result)
+     return true 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 130)
+  def _reduce_45(val, _values, result)
+     return false 
+    result
+  end
+.,.,
+
+# reduce 46 omitted
+
+# reduce 47 omitted
+
+# reduce 48 omitted
+
+# reduce 49 omitted
+
+module_eval(<<'.,.,', 'parser.y', 136)
+  def _reduce_50(val, _values, result)
+     return make_node(:VariableIdentifier, name: val[1], position_source: val[0]) 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 139)
+  def _reduce_51(val, _values, result)
+     return [] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 140)
+  def _reduce_52(val, _values, result)
+     return val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 143)
+  def _reduce_53(val, _values, result)
+     return [val[0]] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 144)
+  def _reduce_54(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 147)
+  def _reduce_55(val, _values, result)
+     return make_node(:InputObject, arguments: [], position_source: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 148)
+  def _reduce_56(val, _values, result)
+     return make_node(:InputObject, arguments: val[1], position_source: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 151)
+  def _reduce_57(val, _values, result)
+     return [val[0]] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 152)
+  def _reduce_58(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 155)
+  def _reduce_59(val, _values, result)
+     return make_node(:Argument, name: val[0], value: val[2], position_source: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 157)
+  def _reduce_60(val, _values, result)
+     return make_node(:Enum, name: val[0], position_source: val[0])
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 160)
+  def _reduce_61(val, _values, result)
+     return [] 
+    result
+  end
+.,.,
+
+# reduce 62 omitted
+
+module_eval(<<'.,.,', 'parser.y', 164)
+  def _reduce_63(val, _values, result)
+     return [val[0]] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 165)
+  def _reduce_64(val, _values, result)
+     val[0] << val[1] 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 167)
+  def _reduce_65(val, _values, result)
+     return make_node(:Directive, name: val[1], arguments: val[2], position_source: val[0]) 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 170)
+  def _reduce_66(val, _values, result)
+     return make_node(:FragmentSpread, name: val[1], directives: val[2], position_source: val[0]) 
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 174)
+  def _reduce_67(val, _values, result)
+          return make_node(:InlineFragment, {
+        type: val[2],
+        directives: val[3],
+        selections: val[4],
+        position_source: val[0]
+      })
+    
+    result
+  end
+.,.,
+
+module_eval(<<'.,.,', 'parser.y', 184)
+  def _reduce_68(val, _values, result)
+          return make_node(:FragmentDefinition, {
+          name:       val[1],
+          type:       val[3],
+          directives: val[4],
+          selections: val[5],
+          position_source: val[0],
+        }
+      )
+    
+    result
+  end
+.,.,
+
+def _reduce_none(val, _values, result)
+  val[0]
+end
+
+    end   # class Parser
+    end   # module Language
+  end   # module GraphQL
